fastapi/requirements.txt
fastapi
pdf2image
python-dotenv
python-multipart
uvicorn
pillow
starlette
boto3
requests
